requirements.txt
# News Intelligence - Dependencies
# Python 3.11+

# Web framework (for local testing)
flask>=3.0.0

# RSS/Feed parsing
feedparser>=6.0.10

# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Google Gemini AI
google-generativeai>=0.3.0

# Date/Time handling
python-dateutil>=2.8.2
pytz>=2024.1

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

